extract REQUEST_STATE to intakeCommon/constants
