+++ client/app/intake/constants.js
@@ -57,13 +57,6 @@ export const ACTIONS = {
   NO_ISSUES_SELECTED_ERROR: 'NO_ISSUES_SELECTED_ERROR'
 };
 
-export const REQUEST_STATE = {
-  NOT_STARTED: 'NOT_STARTED',
-  IN_PROGRESS: 'IN_PROGRESS',
-  SUCCEEDED: 'SUCCEEDED',
-  FAILED: 'FAILED'
-};
-
 export const INTAKE_STATES = {
   NONE: 'NONE',
   STARTED: 'STARTED',
