+++ client/app/intake/pages/rampElection/review.jsx
@@ -9,7 +9,8 @@ import Button from '../../../components/Button';
 import _ from 'lodash';
 import { submitReview } from '../../actions/rampElection';
 import { setReceiptDate, setOptionSelected } from '../../actions/common';
-import { REQUEST_STATE, PAGE_PATHS, INTAKE_STATES, REVIEW_OPTIONS } from '../../constants';
+import { PAGE_PATHS, INTAKE_STATES, REVIEW_OPTIONS } from '../../constants';
+import { REQUEST_STATE } from '../../../intakeCommon/constants';
 import { getIntakeStatus } from '../../selectors';
 import ErrorAlert from '../../components/ErrorAlert';
