+++ client/app/intake/pages/search.jsx
@@ -6,8 +6,8 @@ import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import { Redirect } from 'react-router-dom';
 import { doFileNumberSearch, setFileNumberSearch } from '../actions/common';
-import { REQUEST_STATE, PAGE_PATHS, INTAKE_STATES } from '../constants';
-import { FORM_TYPES } from '../../intakeCommon/constants';
+import { PAGE_PATHS, INTAKE_STATES } from '../constants';
+import { FORM_TYPES, REQUEST_STATE } from '../../intakeCommon/constants';
 import { getIntakeStatus } from '../selectors';
 import _ from 'lodash';
