+++ client/app/intakeEdit/constants.js
@@ -14,10 +14,3 @@ export const ACTIONS = {
   REQUEST_ISSUES_UPDATE_SUCCEED: 'REQUEST_ISSUES_UPDATE_SUCCEED',
   REQUEST_ISSUES_UPDATE_FAIL: 'REQUEST_ISSUES_UPDATE_FAIL'
 };
-
-export const REQUEST_STATE = {
-  NOT_STARTED: 'NOT_STARTED',
-  IN_PROGRESS: 'IN_PROGRESS',
-  SUCCEEDED: 'SUCCEEDED',
-  FAILED: 'FAILED'
-};
